requirements.txt
autopep8==1.6.0
expects==0.9.0
mamba==0.11.2
mysql-connector-python==8.0.28
